Load alda.wasm on alda.io index page
This is just the parser/compiler/OSC generator part. It allows you to run
JavaScript code like `await Alda.toOSCBytes("piano: c")` on alda.io and it will
return a Uint8Array containing the bytes of the OSC bundle transmitted by the
Alda client.

The next step will be to write an in-browser equivalent of the Alda player
process, using the Web Audio API and/or a JS MIDI library.